# hethongvatli1/api-docs.json
{"openapi":"3.1.0","info":{"title":"JSON Web Token Authentication API","description":"Sample JWT authentication service. Demo users: `admin` / `admin123456` and `client` / `client123456`. After sign-in, use **Authorize** and enter `Bearer <token>`.","contact":{"email":"mauriurraco@gmail.com"},"license":{"name":"MIT License","url":"http://opensource.org/licenses/MIT"},"version":"1.0.0"},"servers":[{"url":"http://127.0.0.1:8081","description":"Generated server url"}],"paths":{"/api/v1/users/me":{"get":{"tags":["users"],"summary":"Returns the authenticated user's data","operationId":"whoami","responses":{"200":{"description":"Success","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseUserResponseDTO"}}}},"400":{"description":"Something went wrong","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseUserResponseDTO"}}}},"401":{"description":"Expired or invalid JWT token","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseUserResponseDTO"}}}}},"security":[{"bearerAuth":[]}]},"put":{"tags":["users"],"summary":"Updates current user's username/email","operationId":"updateUserMe","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdateDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseUserResponseDTO"}}}}},"security":[{"bearerAuth":[]}]}},"/api/v1/users/me/profile":{"get":{"tags":["users"],"summary":"Gets current user's profile","operationId":"getMyProfile","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseUserProfileDTO"}}}}},"security":[{"bearerAuth":[]}]},"put":{"tags":["users"],"summary":"Updates current user's profile","operationId":"updateMyProfile","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserProfileUpdateDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseUserProfileDTO"}}}}},"security":[{"bearerAuth":[]}]}},"/api/v1/users/me/password":{"put":{"tags":["users"],"summary":"Changes current user's password","operationId":"changePassword","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePasswordDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}},"security":[{"bearerAuth":[]}]}},"/api/v1/users/admin/users/{id}":{"put":{"tags":["users"],"summary":"Updates user role/email (Admin only)","operationId":"adminUpdateUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminUpdateUserDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseUserResponseDTO"}}}}},"security":[{"bearerAuth":[]}]}},"/api/v1/users/admin/users/{id}/status":{"put":{"tags":["users"],"summary":"Updates user status (Admin only)","operationId":"adminUpdateUserStatus","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserStatusDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseUserResponseDTO"}}}}},"security":[{"bearerAuth":[]}]}},"/api/v1/users/signup":{"post":{"tags":["users"],"summary":"Creates user and returns an access/refresh token pair. Role STUDENT is enforced.","operationId":"signup","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDataDTO"}}},"required":true},"responses":{"200":{"description":"Success","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseAuthResponseDTO"}}}},"400":{"description":"Something went wrong","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseAuthResponseDTO"}}}},"422":{"description":"Username is already in use","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseAuthResponseDTO"}}}}}}},"/api/v1/users/signin":{"post":{"tags":["users"],"summary":"Authenticates user and returns an access/refresh token pair","description":"Rate Limiting is applied to prevent brute-force attacks.","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SigninRequestDTO"}}},"required":true},"responses":{"200":{"description":"Success","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseAuthResponseDTO"}}}},"400":{"description":"Something went wrong","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseAuthResponseDTO"}}}},"422":{"description":"Invalid username/password supplied","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseAuthResponseDTO"}}}}}}},"/api/v1/users/refresh":{"post":{"tags":["users"],"summary":"Exchanges a refresh token for a new access/refresh token pair","description":"Does not require an access token, so it works after the access token has expired. The presented refresh token is consumed and replaced by the one in the response.","operationId":"refresh","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshRequestDTO"}}},"required":true},"responses":{"200":{"description":"New token pair issued","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseAuthResponseDTO"}}}},"400":{"description":"Something went wrong","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseAuthResponseDTO"}}}},"401":{"description":"Expired or invalid refresh token","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseAuthResponseDTO"}}}},"404":{"description":"User no longer exists","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseAuthResponseDTO"}}}}}}},"/api/v1/users/logout":{"post":{"tags":["users"],"summary":"Revokes a refresh token","description":"Idempotent: unknown or already-revoked tokens also return 200. The matching access token stays valid until it expires.","operationId":"logout","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshRequestDTO"}}},"required":true},"responses":{"200":{"description":"Refresh token revoked","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}},"400":{"description":"Something went wrong","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}},"security":[{"bearerAuth":[]}]}},"/api/v1/users/admin/create-user":{"post":{"tags":["users"],"summary":"Creates user with a specific role (Admin only)","operationId":"adminCreateUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminCreateUserDTO"}}},"required":true},"responses":{"200":{"description":"Success","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseUserResponseDTO"}}}},"400":{"description":"Something went wrong","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseUserResponseDTO"}}}},"403":{"description":"Access denied","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseUserResponseDTO"}}}},"422":{"description":"Username is already in use","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseUserResponseDTO"}}}}},"security":[{"bearerAuth":[]}]}},"/api/v1/users/{username}":{"get":{"tags":["users"],"summary":"Returns user by username","operationId":"search","parameters":[{"name":"username","in":"path","description":"Username","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Success","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseUserResponseDTO"}}}},"400":{"description":"Something went wrong","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseUserResponseDTO"}}}},"403":{"description":"Access denied","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseUserResponseDTO"}}}},"404":{"description":"The user doesn't exist","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseUserResponseDTO"}}}}},"security":[{"bearerAuth":[]}]},"delete":{"tags":["users"],"summary":"Deletes user by username","operationId":"delete","parameters":[{"name":"username","in":"path","description":"Username","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Success","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseString"}}}},"400":{"description":"Something went wrong","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseString"}}}},"403":{"description":"Access denied","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseString"}}}},"404":{"description":"The user doesn't exist","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseString"}}}}},"security":[{"bearerAuth":[]}]}},"/api/v1/users/admin/users":{"get":{"tags":["users"],"summary":"Gets all users (Admin only)","operationId":"getAllUsers","parameters":[{"name":"page","in":"query","description":"Zero-based page index (0..N)","required":false,"schema":{"type":"integer","default":0,"minimum":0}},{"name":"size","in":"query","description":"The size of the page to be returned","required":false,"schema":{"type":"integer","default":20,"minimum":1}},{"name":"sort","in":"query","description":"Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.","required":false,"schema":{"type":"array","items":{"type":"string"}}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponsePageUserResponseDTO"}}}}},"security":[{"bearerAuth":[]}]}},"/api/v1/users/admin/users/{id}/profile":{"get":{"tags":["users"],"summary":"Gets user profile by id (Admin only)","operationId":"getUserProfile","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseUserProfileDTO"}}}}},"security":[{"bearerAuth":[]}]}}},"components":{"schemas":{"UserUpdateDTO":{"type":"object","properties":{"username":{"type":"string","description":"Tên đăng nhập mới (tùy chọn)","maxLength":255,"minLength":4},"email":{"type":"string","description":"Email mới (tùy chọn)"}}},"ApiResponseUserResponseDTO":{"type":"object","properties":{"status":{"type":"integer","format":"int32","description":"HTTP Status Code","example":200},"message":{"type":"string","description":"Message describing the result","example":"Success"},"data":{"$ref":"#/components/schemas/UserResponseDTO","description":"The actual data payload"}}},"UserResponseDTO":{"type":"object","properties":{"userId":{"type":"string","format":"uuid"},"username":{"type":"string"},"email":{"type":"string"},"role":{"type":"string","enum":["STUDENT","INSTRUCTOR","TA","ADMIN"]}}},"UserProfileUpdateDTO":{"type":"object","properties":{"fullName":{"type":"string"},"avatarUrl":{"type":"string"},"dateOfBirth":{"type":"string","format":"date"},"gender":{"type":"string","enum":["MALE","FEMALE","OTHER"]},"phone":{"type":"string"},"studentCode":{"type":"string"},"bio":{"type":"string"}}},"ApiResponseUserProfileDTO":{"type":"object","properties":{"status":{"type":"integer","format":"int32","description":"HTTP Status Code","example":200},"message":{"type":"string","description":"Message describing the result","example":"Success"},"data":{"$ref":"#/components/schemas/UserProfileDTO","description":"The actual data payload"}}},"UserProfileDTO":{"type":"object","properties":{"fullName":{"type":"string"},"avatarUrl":{"type":"string"},"dateOfBirth":{"type":"string","format":"date"},"gender":{"type":"string","enum":["MALE","FEMALE","OTHER"]},"phone":{"type":"string"},"studentCode":{"type":"string"},"bio":{"type":"string"}}},"ChangePasswordDTO":{"type":"object","properties":{"oldPassword":{"type":"string","description":"Mật khẩu cũ","minLength":1},"newPassword":{"type":"string","description":"Mật khẩu mới","maxLength":2147483647,"minLength":8}},"required":["newPassword","oldPassword"]},"ApiResponseVoid":{"type":"object","properties":{"status":{"type":"integer","format":"int32","description":"HTTP Status Code","example":200},"message":{"type":"string","description":"Message describing the result","example":"Success"},"data":{"description":"The actual data payload"}}},"AdminUpdateUserDTO":{"type":"object","properties":{"role":{"type":"string","enum":["STUDENT","INSTRUCTOR","TA","ADMIN"]},"email":{"type":"string"}}},"UpdateUserStatusDTO":{"type":"object","properties":{"status":{"type":"string","enum":["ACTIVE","LOCKED"]}},"required":["status"]},"UserDataDTO":{"type":"object","description":"Signup User","properties":{"username":{"type":"string","maxLength":255,"minLength":4},"email":{"type":"string","minLength":1},"password":{"type":"string","maxLength":2147483647,"minLength":8}},"required":["email","password","username"]},"ApiResponseAuthResponseDTO":{"type":"object","properties":{"status":{"type":"integer","format":"int32","description":"HTTP Status Code","example":200},"message":{"type":"string","description":"Message describing the result","example":"Success"},"data":{"$ref":"#/components/schemas/AuthResponseDTO","description":"The actual data payload"}}},"AuthResponseDTO":{"type":"object","properties":{"accessToken":{"type":"string","description":"Short-lived JWT to send as 'Authorization: Bearer <token>'"},"refreshToken":{"type":"string","description":"Long-lived opaque token used to obtain a new access token"},"tokenType":{"type":"string","description":"Authentication scheme the access token is used with","example":"Bearer"},"expiresIn":{"type":"integer","format":"int64","description":"Access token lifetime in seconds","example":3600},"refreshExpiresIn":{"type":"integer","format":"int64","description":"Refresh token lifetime in seconds","example":604800}}},"SigninRequestDTO":{"type":"object","description":"Signin Credentials","properties":{"username":{"type":"string","description":"Username for authentication","example":"admin","minLength":1},"password":{"type":"string","description":"Password for authentication","example":"admin123456","minLength":1}},"required":["password","username"]},"RefreshRequestDTO":{"type":"object","properties":{"refreshToken":{"type":"string","description":"Refresh token used to obtain a new access token","minLength":1}},"required":["refreshToken"]},"AdminCreateUserDTO":{"type":"object","description":"Admin Create User","properties":{"username":{"type":"string","description":"Username of the new user","example":"newinstructor","maxLength":255,"minLength":4},"email":{"type":"string","description":"Email of the new user","example":"instructor@example.com","minLength":1},"password":{"type":"string","description":"Password of the new user","example":"password123","maxLength":2147483647,"minLength":8},"role":{"type":"string","description":"Role assigned by Admin","enum":["STUDENT","INSTRUCTOR","TA","ADMIN"],"example":"INSTRUCTOR"}},"required":["email","password","role","username"]},"ApiResponsePageUserResponseDTO":{"type":"object","properties":{"status":{"type":"integer","format":"int32","description":"HTTP Status Code","example":200},"message":{"type":"string","description":"Message describing the result","example":"Success"},"data":{"$ref":"#/components/schemas/PageUserResponseDTO","description":"The actual data payload"}}},"PageUserResponseDTO":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/UserResponseDTO"}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"first":{"type":"boolean"},"last":{"type":"boolean"},"numberOfElements":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"empty":{"type":"boolean"}}},"PageableObject":{"type":"object","properties":{"offset":{"type":"integer","format":"int64"},"sort":{"$ref":"#/components/schemas/SortObject"},"paged":{"type":"boolean"},"pageNumber":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"unpaged":{"type":"boolean"}}},"SortObject":{"type":"object","properties":{"empty":{"type":"boolean"},"sorted":{"type":"boolean"},"unsorted":{"type":"boolean"}}},"ApiResponseString":{"type":"object","properties":{"status":{"type":"integer","format":"int32","description":"HTTP Status Code","example":200},"message":{"type":"string","description":"Message describing the result","example":"Success"},"data":{"type":"string","description":"The actual data payload"}}}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}